add /logs endpoint
